backend/pyproject.toml
[project]
name = "rqg"
version = "0.1.0"
description = ""
authors = [
    { name = "nico-vromans", email = "VromansN+GitHub@pm.me" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.4,<6.0.0)",
    "django-configurations[cache,database,email,search] (>=2.5.1,<3.0.0) ; python_version == \"3.13\"",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "markdown (>=3.7,<4.0)",
    "django-filter (>=24.3,<25.0)",
    "django-unfold (>=0.44.0,<0.45.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.10.5,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
